requirements.txt
flask
tabula-py
flask-bootstrap
xlrd